Fix the version mismatch in the package.json
