+++ asset/package.json
@@ -14,7 +14,7 @@
         "build:watch": "yarn build --watch"
     },
     "dependencies": {
-        "@terascope/file-asset-apis": "^0.1.0",
+        "@terascope/file-asset-apis": "^0.2.0",
         "@terascope/job-components": "^0.46.2",
         "csvtojson": "^2.0.10",
         "fs-extra": "^9.1.0",
