+++ asset/tsconfig.build.json
@@ -14,8 +14,7 @@
         "noImplicitReturns": true,
         "preserveConstEnums": true,
         "esModuleInterop": true,
-        "resolveJsonModule": true,
-        "suppressImplicitAnyIndexErrors": true
+        "resolveJsonModule": true
     },
     "include": [
         "src/**/*"
