+++ package.json
@@ -26,16 +26,14 @@
     "dependencies": {},
     "devDependencies": {
         "@terascope/eslint-config": "^0.5.3",
-        "@terascope/file-asset-apis": "^0.1.0",
+        "@terascope/file-asset-apis": "^0.2.0",
         "@terascope/job-components": "^0.46.2",
-        "@types/bluebird": "^3.5.33",
         "@types/fs-extra": "^9.0.7",
         "@types/jest": "^26.0.19",
         "@types/json2csv": "^5.0.1",
         "@types/node": "^14.14.16",
         "@types/node-gzip": "^1.1.0",
         "aws-sdk": "^2.843.0",
-        "bluebird": "^3.7.2",
         "eslint": "^7.20.0",
         "fs-extra": "^9.1.0",
         "jest": "^26.6.3",
