+++ packages/file-asset-apis/src/index.ts
@@ -1,5 +1,5 @@
+export * from './base';
 export * from './file-api';
 export * from './hdfs-api';
-export * from './s3-api';
-export * from './base';
 export * from './interfaces';
+export * from './s3-api';
