+++ tsconfig.json
@@ -13,7 +13,6 @@
         "esModuleInterop": true,
         "resolveJsonModule": true,
         "forceConsistentCasingInFileNames": true,
-        "suppressImplicitAnyIndexErrors": true,
         "composite": true,
         "declaration": true,
         "declarationMap": true,
